refactor(nourish): extract runScoringPipeline from /api/nourish
Commit 9a of PR 4. Pure refactor — no user-visible behavior change.
Extracts the LLM scoring pipeline (prompt build → OpenAI call →
parse → clamp/validate → assemble NourishScores) from
/api/nourish/+server.ts into a new server-side module
(src/lib/nourish/scoringEngine.server.ts). Prepares the ground for
commit 9b's admin rescore endpoint, which will drive the same
pipeline without duplicating the implementation.

The new module owns:
- NOURISH_PROMPT template
- clampScore, validateLabel, validateImprovements,
  validateIngredientSignals (module-private — runScoringPipeline
  is the only export callers need)
- runScoringPipeline(openAiKey, { title, ingredients, tags, servings })
  returning a discriminated { ok: true, scores, improvements,
  ingredientSignals } | { ok: false, status, error } result

Kept in the endpoint:
- OPENAI_API_KEY boundary check
- Request-body shape validation
- Membership gate (requireMembership)
- HEX_64_RE / recipePubkey / recipeDTag / contentHash gates for
  pantry publish
- publishNourishEvent call
- Final JSON response assembly (NOURISH_PROMPT_VERSION, contentHash
  echo, createdAt stamp)

Shape preservation — audited every response path:
- 4 validation error paths (OPENAI key, ingredients, eventId,
  membership): identical body + status
- 3 pipeline error paths (OpenAI fetch, empty content, parse
  failure): all route through pipelineResult.ok === false with the
  exact error strings and status 500 from pre-refactor
- Outer catch: unchanged
- Success response field order preserved: success, scores,
  improvements, ingredient_signals, promptVersion, contentHash,
  createdAt
- scores object shape: identical (gut/protein/realFood/overall
  each with score/label/reason; summary; cacheVersion)
- improvements: validateImprovements does filter+slice(0,5)
  matching pre-refactor behavior
- ingredient_signals (snake_case on the wire): helper returns
  ingredientSignals (camelCase); endpoint aliases to preserve the
  snake_case API field name. Element shape (name lowercased-and-
  trimmed, signals, contribution allowlist) identical.

Non-deterministic LLM score values naturally differ between calls —
that's inherent to the model, not the refactor. The verification
gate per Phase 2 §10 is shape-byte-for-byte, which this preserves.

Verified:
- pnpm run check: 4 errors / 127 warnings (baseline preserved)
- pnpm run build: passes

Commit 9b will add /api/admin/nourish/rescore using the extracted
runScoringPipeline plus a new publishNourishEvent updatedAt override.
