+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "zap.cooking",
   "license": "MIT",
-  "version": "4.2.242",
+  "version": "4.2.243",
   "private": true,
   "scripts": {
     "dev": "vite dev",
